usecombis: alter templates to include dependencies
Some USE instances conflict between a target package and it's
dependencies (such as a package having USE=qt but depending on
atom[-qt]). This caters to that, and also introduces more thorough
testing, but making use of package.use and package.env. Note that this
is intended for minimal environments dedecated to testing but can be
used for other normal environments (with the caveat that deps will be
changing while runtesting).